Implement IMU AHRS filters to obtain absolute orientation (#3)
AHRS = attitude heading reference system

This PR adds the Madgwick and Mahony filters to the library crate, as
ported from
https://github.com/drcpattison/ICM-20948/blob/master/src/AHRSAlgorithms.cpp.

Both filters work by using 9DOF IMU measurements (gyro + accel + mag) to
compute a the orientation of the IMU represented by a quaternion. The
quaternion is converted to Euler angles (roll + pitch + yaw) for use in
our application.

Testing showed that the Madgwick filter is unusable as the filtered yaw
values are very inaccurate. The Mahony filter is able to provide much
more accurate yaw readings, but with the issue that the yaw error
accumulates over time. After optimizing the gain values with Ki = 0 and
Kp = 0, this error accumulates at ~0.2 deg/s, which I think should be
good enough for what we're using it for as long as we reset the filter
from time to time.

Edit: idk how the math actually works but it just does


Update: I discovered that if I account for the gyro bias before updating
the Madgwick or Mahony filter, I get fairly accurate filter results. I'm
not sure why the filters don't account for this bias internally, but
this fix seems to give decent orientation readings. Both the Madgwick
and Mahony filters are usable for our final app.